ETL scripts are now stopping on errors. Multiple etl issues fixed.

+++ etl-vz/etl/stage/reference/geotel/geotel_etl.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+set -e;
+
 
 PSQL="${PGBIN}/psql -v ON_ERROR_STOP=1"
 SHP2PGSQL=${PGBIN}/shp2pgsql
@@ -19,20 +21,27 @@ fiber_state_array_len=${#FIBER_STATE_ARRAY[@]}
 # If there is only one state, download the file and create the table
 if [ ${fiber_state_array_len} == 1 ]; then
 	rm -f ${TMPDIR}/*.*
-	aws s3 cp s3://public.aro/geotel/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip
+
+	if [ ! -f $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip ]; then
+		aws s3 cp s3://public.aro/geotel/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip
+	fi
 	$UNZIPTOOL geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip -d ${TMPDIR}
 	${SHP2PGSQL} -c -s 4326 -g the_geom -t 2D -W "latin1" /$TMPDIR/geotel_fiber_${FIBER_STATE_ARRAY[0]}.dbf geotel.fiber_plant | ${PSQL}
 # If there are two or more states, download the first file, create the table, then loop through the rest and append
 elif [ ${fiber_state_array_len} > 1 ]; then
 	rm -f ${TMPDIR}/*.*
-	aws s3 cp s3://public.aro/geotel/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip
+        if [ ! -f $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip ]; then
+		aws s3 cp s3://public.aro/geotel/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip
+	fi
 	$UNZIPTOOL geotel_fiber_${FIBER_STATE_ARRAY[0]}.zip -d ${TMPDIR}
 	${SHP2PGSQL} -c -s 4326 -g the_geom -t 2D -W "latin1" /$TMPDIR/geotel_fiber_${FIBER_STATE_ARRAY[0]}.dbf geotel.fiber_plant | ${PSQL}
 
 	for ((i=1; i<$fiber_state_array_len; i++ ));
 	do
 		rm -f ${TMPDIR}/*.*
-		aws s3 cp s3://public.aro/geotel/geotel_fiber_${FIBER_STATE_ARRAY[i]}.zip $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[i]}.zip
+	        if [ ! -f $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[i]}.zip ]; then
+			aws s3 cp s3://public.aro/geotel/geotel_fiber_${FIBER_STATE_ARRAY[i]}.zip $GISROOT/geotel_fiber_${FIBER_STATE_ARRAY[i]}.zip
+		fi
 		$UNZIPTOOL geotel_fiber_${FIBER_STATE_ARRAY[i]}.zip -d ${TMPDIR}
 		${SHP2PGSQL} -a -s 4326 -g the_geom -t 2D -W "latin1" /$TMPDIR/geotel_fiber_${FIBER_STATE_ARRAY[i]}.dbf geotel.fiber_plant | ${PSQL}
 	done
@@ -46,20 +55,26 @@ wirecenter_state_array_len=${#WIRECENTER_STATE_ARRAY[@]}
 # If there is only one state, download the file and create the table
 if [ ${wirecenter_state_array_len} == 1 ]; then
 	rm -f ${TMPDIR}/*.*
-	aws s3 cp s3://public.aro/geotel/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip
+        if [ ! -f $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip ]; then
+		aws s3 cp s3://public.aro/geotel/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip
+	fi
 	$UNZIPTOOL geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip -d ${TMPDIR}
 	${SHP2PGSQL} -c -s 4326 -g the_geom -t 2D -W "latin1" /$TMPDIR/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.dbf geotel.wirecenters | ${PSQL}
 # If there are two or more states, download the first file, create the table, then loop through the rest and append
 elif [ ${wirecenter_state_array_len} > 1 ]; then
 	rm -f ${TMPDIR}/*.*
-	aws s3 cp s3://public.aro/geotel/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip
+        if [ ! -f $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip ]; then
+		aws s3 cp s3://public.aro/geotel/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip
+	fi
 	$UNZIPTOOL geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.zip -d ${TMPDIR}
 	${SHP2PGSQL} -c -s 4326 -g the_geom -t 2D -W "latin1" /$TMPDIR/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[0]}.dbf geotel.wirecenters | ${PSQL}
 
 	for ((i=1; i<$wirecenter_state_array_len; i++ ));
 	do
 		rm -f ${TMPDIR}/*.*
-		aws s3 cp s3://public.aro/geotel/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[i]}.zip $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[i]}.zip
+	        if [ ! -f $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[i]}.zip ]; then
+			aws s3 cp s3://public.aro/geotel/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[i]}.zip $GISROOT/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[i]}.zip
+		fi
 		$UNZIPTOOL geotel_wirecenters_${WIRECENTER_STATE_ARRAY[i]}.zip -d ${TMPDIR}
 		${SHP2PGSQL} -a -s 4326 -g the_geom -t 2D -W "latin1" /$TMPDIR/geotel_wirecenters_${WIRECENTER_STATE_ARRAY[i]}.dbf geotel.wirecenters | ${PSQL}
 	done
